refactor: #347
TicketApiClient가 TicketClient를 구현하도록 수정
